Fixed timing being incorrectly passed to pretender
