+++ lib/mock-server/pretender-config.js
@@ -2,6 +2,7 @@ import "@miragejs/pretender-node-polyfill/before";
 import Pretender from "pretender";
 import "@miragejs/pretender-node-polyfill/after";
 import assert from "../assert";
+import assign from "lodash.assign";
 
 /**
   Mirage Interceptor Class
@@ -47,10 +48,64 @@ const defaultPassthroughs = [
   },
 ];
 
+const defaultRouteOptions = {
+  coalesce: false,
+  timing: undefined,
+};
+
+/**
+ * Determine if the object contains a valid option.
+ *
+ * @method isOption
+ * @param {Object} option An object with one option value pair.
+ * @return {Boolean} True if option is a valid option, false otherwise.
+ * @private
+ */
+function isOption(option) {
+  if (!option || typeof option !== "object") {
+    return false;
+  }
+
+  let allOptions = Object.keys(defaultRouteOptions);
+  let optionKeys = Object.keys(option);
+  for (let i = 0; i < optionKeys.length; i++) {
+    let key = optionKeys[i];
+    if (allOptions.indexOf(key) > -1) {
+      return true;
+    }
+  }
+  return false;
+}
+
 /**
  @hide
  */
 export { defaultPassthroughs };
+
+/**
+ * Extract arguments for a route.
+ *
+ * @method extractRouteArguments
+ * @param {Array} args Of the form [options], [object, code], [function, code]
+ * [shorthand, options], [shorthand, code, options]
+ * @return {Array} [handler (i.e. the function, object or shorthand), code,
+ * options].
+ */
+function extractRouteArguments(args) {
+  let [lastArg] = args.splice(-1);
+  if (isOption(lastArg)) {
+    lastArg = assign({}, defaultRouteOptions, lastArg);
+  } else {
+    args.push(lastArg);
+    lastArg = defaultRouteOptions;
+  }
+  let t = 2 - args.length;
+  while (t-- > 0) {
+    args.push(undefined);
+  }
+  args.push(lastArg);
+  return args;
+}
 
 export default class PretenderConfig {
   urlPrefix;
@@ -106,10 +161,17 @@ export default class PretenderConfig {
       ["options"],
     ].forEach(([verb, alias]) => {
       this[verb] = (path, ...args) => {
-        let handler = mirageServer.registerRouteHandler(verb, path, args);
+        let [rawHandler, customizedCode, options] = extractRouteArguments(args);
+        let handler = mirageServer.registerRouteHandler(
+          verb,
+          path,
+          rawHandler,
+          customizedCode,
+          options
+        );
         let fullPath = this._getFullPath(path);
         let timing =
-          config.timing !== undefined ? config.timing : () => this.timing;
+          options.timing !== undefined ? options.timing : () => this.timing;
         return this.pretender?.[verb](fullPath, handler, timing);
       };
