+++ lib/server.js
@@ -16,62 +16,7 @@ import PretenderConfig from "./mock-server/pretender-config";
 
 const isPluralForModelCache = {};
 
-const defaultRouteOptions = {
-  coalesce: false,
-  timing: undefined,
-};
-
 const defaultInflector = { singularize, pluralize };
-
-/**
- * Determine if the object contains a valid option.
- *
- * @method isOption
- * @param {Object} option An object with one option value pair.
- * @return {Boolean} True if option is a valid option, false otherwise.
- * @private
- */
-function isOption(option) {
-  if (!option || typeof option !== "object") {
-    return false;
-  }
-
-  let allOptions = Object.keys(defaultRouteOptions);
-  let optionKeys = Object.keys(option);
-  for (let i = 0; i < optionKeys.length; i++) {
-    let key = optionKeys[i];
-    if (allOptions.indexOf(key) > -1) {
-      return true;
-    }
-  }
-  return false;
-}
-
-/**
- * Extract arguments for a route.
- *
- * @method extractRouteArguments
- * @param {Array} args Of the form [options], [object, code], [function, code]
- * [shorthand, options], [shorthand, code, options]
- * @return {Array} [handler (i.e. the function, object or shorthand), code,
- * options].
- * @private
- */
-function extractRouteArguments(args) {
-  let [lastArg] = args.splice(-1);
-  if (isOption(lastArg)) {
-    lastArg = assign({}, defaultRouteOptions, lastArg);
-  } else {
-    args.push(lastArg);
-    lastArg = defaultRouteOptions;
-  }
-  let t = 2 - args.length;
-  while (t-- > 0) {
-    args.push(undefined);
-  }
-  args.push(lastArg);
-  return args;
-}
 
 /**
  * Creates a Server
@@ -768,18 +713,7 @@ export default class Server {
     }
   }
 
-  registerRouteHandler(verb, path, args) {
-    let [rawHandler, customizedCode, options] = extractRouteArguments(args);
-    return this._registerRouteHandler(
-      verb,
-      path,
-      rawHandler,
-      customizedCode,
-      options
-    );
-  }
-
-  _registerRouteHandler(verb, path, rawHandler, customizedCode, options) {
+  registerRouteHandler(verb, path, rawHandler, customizedCode, options) {
     let routeHandler = this._container.create("RouteHandler", {
       schema: this.schema,
       verb,
